Added transactions with lock in photosRepository. Added ability to cancel deletion of failed photo (via snackbar)

+++ app/src/main/java/com/kirakishou/photoexchange/helper/database/repository/PhotosRepository.kt
@@ -3,6 +3,7 @@ package com.kirakishou.photoexchange.helper.database.repository
 import com.kirakishou.photoexchange.helper.database.MyDatabase
 import com.kirakishou.photoexchange.helper.database.entity.MyPhotoEntity
 import com.kirakishou.photoexchange.helper.database.entity.TempFileEntity
+import com.kirakishou.photoexchange.helper.database.isFail
 import com.kirakishou.photoexchange.helper.database.mapper.MyPhotoEntityMapper
 import com.kirakishou.photoexchange.mvp.model.MyPhoto
 import com.kirakishou.photoexchange.mvp.model.PhotoState
@@ -21,114 +22,145 @@ open class PhotosRepository(
     private val tag = "PhotosRepository"
     private val myPhotoDao = database.myPhotoDao()
     private val tempFileDao = database.tempFileDao()
-    private val lock = ReentrantLock()
 
     init {
         createTempFilesDirIfNotExists()
     }
 
     fun saveTakenPhoto(file: File): MyPhoto {
-        val tempFileEntity = TempFileEntity.create(file.absolutePath)
-        val tempFileId = tempFileDao.insert(tempFileEntity)
-        if (tempFileId <= 0L) {
-            return MyPhoto.empty()
+        var photo = MyPhoto.empty()
+
+        database.transactional {
+            val tempFileEntity = TempFileEntity.create(file.absolutePath)
+            val tempFileId = tempFileDao.insert(tempFileEntity)
+            if (tempFileId <= 0L) {
+                photo = MyPhoto.empty()
+                return@transactional false
+            }
+
+            val myPhotoEntity = MyPhotoEntity.create(tempFileId, false)
+            val myPhotoId = myPhotoDao.insert(myPhotoEntity)
+
+            myPhotoEntity.id = myPhotoId
+            photo = MyPhotoEntityMapper.toMyPhoto(myPhotoEntity, tempFileEntity)
+
+            return@transactional true
         }
 
-        val myPhotoEntity = MyPhotoEntity.create(tempFileId, false)
-        val myPhotoId = myPhotoDao.insert(myPhotoEntity)
-
-        myPhotoEntity.id = myPhotoId
-        return MyPhotoEntityMapper.toMyPhoto(myPhotoEntity, tempFileEntity)
+        return photo
     }
 
     fun updateSetPhotoName(photoId: Long, photoName: String): Boolean {
-        return myPhotoDao.updateSetPhotoName(photoId, photoName) == 1
+        return database.transactional {
+            myPhotoDao.updateSetPhotoName(photoId, photoName) == 1
+        }
     }
 
     fun updateSetTempFileId(photoId: Long, newTempFileId: Long?): Boolean {
-        return myPhotoDao.updateSetTempFileId(photoId, newTempFileId) == 1
+        return database.transactional {
+            myPhotoDao.updateSetTempFileId(photoId, newTempFileId) == 1
+        }
     }
 
     fun updatePhotoState(photoId: Long, newPhotoState: PhotoState): Boolean {
-        return myPhotoDao.updateSetNewPhotoState(photoId, newPhotoState) == 1
+        return database.transactional {
+            myPhotoDao.updateSetNewPhotoState(photoId, newPhotoState) == 1
+        }
     }
 
     fun updatePhotoState(photoName: String, newPhotoState: PhotoState): Boolean {
-        val photoId = myPhotoDao.findByName(photoName)?.id ?: return false
-        return myPhotoDao.updateSetNewPhotoState(photoId, newPhotoState) == 1
+        return database.transactional {
+            val photoId = myPhotoDao.findByName(photoName)?.id
+                ?: return@transactional false
+
+            return@transactional myPhotoDao.updateSetNewPhotoState(photoId, newPhotoState) == 1
+        }
     }
 
     fun updateMakePhotoPublic(takenPhotoId: Long): Boolean {
-        return myPhotoDao.updateSetPhotoPublic(takenPhotoId) == 1
-    }
-
-    fun deleteMyPhoto(myPhoto: MyPhoto?): Boolean {
-        if (myPhoto == null) {
-            return true
+        return database.transactional {
+            return myPhotoDao.updateSetPhotoPublic(takenPhotoId) == 1
         }
-
-        if (myPhoto.isEmpty()) {
-            return true
-        }
-
-        return deleteById(myPhoto.id)
     }
 
     fun createFile(): File {
-        filesDir
         val fullPathFile = File(filesDir)
         return File.createTempFile("file_", ".tmp", fullPathFile)
     }
 
     fun findById(id: Long): MyPhoto {
-        val myPhotoEntity = myPhotoDao.findById(id) ?: MyPhotoEntity.empty()
-        val tempFileEntity = findTempFileById(id)
+        var photo = MyPhoto.empty()
 
-        return MyPhotoEntityMapper.toMyPhoto(myPhotoEntity, tempFileEntity)
+        database.transactional {
+            val myPhotoEntity = myPhotoDao.findById(id) ?: MyPhotoEntity.empty()
+            val tempFileEntity = findTempFileById(id)
+
+            photo = MyPhotoEntityMapper.toMyPhoto(myPhotoEntity, tempFileEntity)
+            return@transactional true
+        }
+
+        return photo
     }
 
     fun findAll(): List<MyPhoto> {
-        val allMyPhotos = arrayListOf<MyPhoto>()
-        val allMyPhotoEntities = myPhotoDao.findAll()
+        var photos = emptyList<MyPhoto>()
 
-        for (myPhotoEntity in allMyPhotoEntities) {
-            myPhotoEntity.id?.let { myPhotoId ->
-                val tempFile = findTempFileById(myPhotoId)
+        database.transactional {
+            val allMyPhotos = arrayListOf<MyPhoto>()
+            val allMyPhotoEntities = myPhotoDao.findAll()
 
-                MyPhotoEntityMapper.toMyPhoto(myPhotoEntity, tempFile).let { myPhoto ->
-                    allMyPhotos += myPhoto
+            for (myPhotoEntity in allMyPhotoEntities) {
+                myPhotoEntity.id?.let { myPhotoId ->
+                    val tempFile = findTempFileById(myPhotoId)
+
+                    MyPhotoEntityMapper.toMyPhoto(myPhotoEntity, tempFile).let { myPhoto ->
+                        allMyPhotos += myPhoto
+                    }
                 }
             }
+
+            photos = allMyPhotos
+            return@transactional true
         }
 
-        return allMyPhotos
+        return photos
     }
 
     fun countAllByState(state: PhotoState): Int {
-        return myPhotoDao.countAllByState(state).toInt()
+        var count = 0
+
+        database.transactional {
+            count = myPhotoDao.countAllByState(state).toInt()
+            return@transactional true
+        }
+
+        return count
     }
 
     fun countAllByStates(states: Array<PhotoState>): Int {
-        return myPhotoDao.countAllByStates(states)
+        var count = 0
+
+        database.transactional {
+            count = myPhotoDao.countAllByStates(states)
+            return@transactional true
+        }
+
+        return count
     }
 
     fun findPhotosByStateAndUpdateState(oldState: PhotoState, newState: PhotoState): List<MyPhoto> {
-        return lock.withLock {
-            var photos = mutableListOf<MyPhotoEntity>()
+        var resultPhotos = emptyList<MyPhoto>()
 
-            database.transactional {
-                photos = myPhotoDao.findOnePhotoWithState(oldState)
-                if (photos.isEmpty()) {
+        database.transactional {
+            val photos = myPhotoDao.findOnePhotoWithState(oldState)
+            if (photos.isEmpty()) {
+                return@transactional false
+            }
+
+            for (photo in photos) {
+                if (myPhotoDao.updateSetNewPhotoState(photo.id!!, newState) != 1) {
                     return@transactional false
                 }
-
-                for (photo in photos) {
-                    if (myPhotoDao.updateSetNewPhotoState(photo.id!!, newState) != 1) {
-                        return@transactional false
-                    }
-                }
-
-                return@transactional true
             }
 
             for (photo in photos) {
@@ -137,21 +169,27 @@ open class PhotosRepository(
                     return@let MyPhotoEntityMapper.toMyPhoto(photo, tempFileEntity)
                 }
             }
-            return@withLock photos.asSequence()
+
+            resultPhotos = photos.asSequence()
                 .filter { it.id != null }
                 .map { photo ->
                     val tempFileEntity = findTempFileById(photo.id!!)
                     MyPhotoEntityMapper.toMyPhoto(photo, tempFileEntity)
                 }
                 .toList()
+
+            return@transactional true
         }
+
+        return resultPhotos
     }
 
     fun findAllByState(state: PhotoState): List<MyPhoto> {
-        val allPhotoReadyToUploading = myPhotoDao.findAllWithState(state)
         val resultList = mutableListOf<MyPhoto>()
 
         database.transactional {
+            val allPhotoReadyToUploading = myPhotoDao.findAllWithState(state)
+
             for (photo in allPhotoReadyToUploading) {
                 val tempFileEntity = findTempFileById(photo.id!!)
                 resultList += MyPhotoEntityMapper.toMyPhoto(photo, tempFileEntity)
@@ -164,28 +202,36 @@ open class PhotosRepository(
     }
 
     fun findByPhotoIdByName(photoName: String): Long {
-        return myPhotoDao.findPhotoIdByName(photoName) ?: -1L
-    }
+        var photoId = -1L
 
-    private fun deleteById(id: Long): Boolean {
-        val tempFileEntity = findTempFileById(id)
-        var fileDeleteResult = false
-
-        if (!tempFileEntity.isEmpty()) {
-            val photoFile = File(tempFileEntity.filePath)
-            if (photoFile.exists()) {
-                fileDeleteResult = photoFile.delete()
-            }
+        database.transactional {
+            photoId = myPhotoDao.findPhotoIdByName(photoName) ?: -1L
+            return@transactional true
         }
 
-        val myPhotoDeleteResult = myPhotoDao.deleteById(id) > 0
-        val tempFileDeleteResult = deleteTempFileById(id)
-
-        return myPhotoDeleteResult && tempFileDeleteResult && fileDeleteResult
+        return photoId
     }
 
-    fun deletePhotoById(photoId: Long) {
-        myPhotoDao.deleteById(photoId)
+    fun deleteMyPhoto(myPhoto: MyPhoto?): Boolean {
+        if (myPhoto == null) {
+            return true
+        }
+
+        if (myPhoto.isEmpty()) {
+            return true
+        }
+
+        return deletePhotoById(myPhoto.id)
+    }
+
+    fun deletePhotoById(photoId: Long): Boolean {
+        return database.transactional {
+            if (myPhotoDao.deleteById(photoId).isFail()) {
+                return@transactional false
+            }
+
+            return@transactional deleteTempFileById(photoId)
+        }
     }
 
     fun deleteAllWithState(photoState: PhotoState): Boolean {
@@ -193,7 +239,7 @@ open class PhotosRepository(
             val myPhotosList = myPhotoDao.findAllWithState(photoState)
 
             for (myPhoto in myPhotosList) {
-                if (!deleteById(myPhoto.id!!)) {
+                if (!deletePhotoById(myPhoto.id!!)) {
                     return@transactional false
                 }
             }
@@ -202,16 +248,29 @@ open class PhotosRepository(
         }
     }
 
-    private fun findTempFileById(id: Long): TempFileEntity {
-        return tempFileDao.findById(id) ?: TempFileEntity.empty()
+    fun deleteTempFileById(id: Long): Boolean {
+        return database.transactional {
+            val tempFileEntity = tempFileDao.findById(id)
+                ?: return@transactional true
+
+            if (tempFileDao.deleteById(id).isFail()) {
+                return@transactional false
+            }
+
+            deleteFileIfExists(File(tempFileEntity.filePath))
+            return@transactional true
+        }
     }
 
-    fun deleteTempFileById(id: Long): Boolean {
-        val tempFileEntity = tempFileDao.findById(id)
-            ?: return true
+    private fun findTempFileById(id: Long): TempFileEntity {
+        var tempFileEntity = TempFileEntity.empty()
 
-        deleteFileIfExists(File(tempFileEntity.filePath))
-        return tempFileDao.deleteById(id) > 0
+        database.transactional {
+            tempFileEntity = tempFileDao.findById(id) ?: TempFileEntity.empty()
+            return@transactional true
+        }
+
+        return tempFileEntity
     }
 
     private fun createTempFilesDirIfNotExists() {
@@ -239,15 +298,19 @@ open class PhotosRepository(
         val directory = File(filesDir)
         val filePaths = directory.listFiles()
 
-        for (filePath in filePaths) {
-            if (filePath.name.startsWith(".")) {
-                continue
+        database.transactional {
+            for (filePath in filePaths) {
+                if (filePath.name.startsWith(".")) {
+                    continue
+                }
+
+                val found = tempFileDao.findByFilePath(filePath.absolutePath) != null
+                if (!found) {
+                    deleteFileIfExists(filePath)
+                }
             }
 
-            val found = tempFileDao.findByFilePath(filePath.absolutePath) != null
-            if (!found) {
-                deleteFileIfExists(filePath)
-            }
+            return@transactional true
         }
     }
 }
